adding task description and adding page for task details and editing tasks
