+++ client/src/components/css/TaskDetail.css
@@ -1,75 +1,94 @@
 .task-detail {
-  display: grid;
-  grid-template-columns: 1fr;
-  justify-content: center;
-  align-items: center;
-  padding: 20px;
-  background-color: #f9f9f9;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  width: 40rem;
-  margin: 0 auto;
-  height: auto;
+  max-width: 800px;
+  margin: 2rem auto;
+  padding: 2rem 2.5rem;
+  background: #fff;
+  border-radius: 16px;
+  box-shadow: 0 2px 12px rgba(44, 62, 80, 0.07);
+  border: 1.5px solid #e5e4e4;
+  font-family: "Inter", sans-serif;
+  color: #222;
 }
 
-.task-detail label {
-  display: flex;
-  flex-direction: column;
-  font-weight: 500;
+.task-detail h2 {
+  font-size: 1.6rem;
+  font-weight: 700;
+  margin-bottom: 1.5rem;
+  color: #258bff;
+  border-bottom: 2px solid #e5e4e4;
+  padding-bottom: 0.5rem;
+}
+
+.task-detail p {
+  font-size: 1rem;
+  margin: 0.8rem 0;
+  line-height: 1.5;
   color: #333;
 }
 
-.task-detail input[type="text"],
-.task-detail select {
-  padding: 0.5rem;
-  border-radius: 4px;
-  border: 1px solid #ccc;
-  font-size: 1rem;
-  margin-top: 0.3rem;
-}
-
-.task-detail select[multiple] {
-  height: 6rem;
-}
-
-.task-detail button[type="submit"] {
-  padding: 0.6rem 1rem;
-  background-color: #4caf50;
-  color: white;
-  font-weight: bold;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.task-detail button[type="submit"]:disabled {
-  background-color: #a5d6a7;
-  cursor: not-allowed;
-}
-
-.task-detail button:last-of-type {
-  padding: 0.6rem 1rem;
-  background-color: #e53935;
-  color: white;
-  font-weight: bold;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
+.task-detail strong {
+  font-weight: 600;
+  color: #222;
+  margin-right: 0.4rem;
 }
 
 .task-detail .label {
-  display: inline-block;
-  background-color: #e0e0e0;
-  color: #333;
-  padding: 0.3rem 0.6rem;
-  border-radius: 4px;
-  margin-right: 0.5rem;
-  font-size: 0.9rem;
+  margin-right: 0.4rem;
+  margin-bottom: 0.4rem;
+  display: inline-flex;
+  align-items: center;
 }
 
-.task-detail a {
-  margin-top: 1rem;
+/* Status styling */
+.task-status {
+  display: inline-block;
+  font-weight: 700;
+  padding: 0.25rem 0.6rem;
+  border-radius: 6px;
+  font-size: 0.9rem;
+  margin-left: 0.4rem;
+}
+
+/* Status colors */
+.task-status.todo {
+  color: #258bff;
+  background: #f5f7fa;
+  border-left: 4px solid #258bff;
+}
+
+.task-status.in-progress {
+  color: #f6b100;
+  background: #fffbe6;
+  border-left: 4px solid #f6b100;
+}
+
+.task-status.ready-for-review {
+  color: #1db96d;
+  background: #eafaf1;
+  border-left: 4px solid #1db954;
+}
+
+.task-status.done {
+  color: #e53935;
+  background: #fff0f0;
+  border-left: 4px solid #e53935;
+}
+
+.task-detail__button {
+  display: inline-block;
+  margin-top: 1.8rem;
+  padding: 0.6rem 1.2rem;
+  background-color: #258bff;
+  color: #fff;
+  font-weight: 600;
+  border-radius: 8px;
   text-decoration: none;
-  color: #1976d2;
-  font-weight: bold;
+  transition: background 0.2s, box-shadow 0.2s;
+  margin-right: 0.5rem;
+  width: fit-content;
+}
+
+.task-detail__button:hover {
+  background-color: #1474d9;
+  box-shadow: 0 4px 14px rgba(37, 139, 255, 0.2);
 }
