+++ server/src/api/task/content-types/task/schema.json
@@ -37,6 +37,9 @@
       "relation": "manyToMany",
       "target": "api::label.label",
       "mappedBy": "tasks"
+    },
+    "description": {
+      "type": "text"
     }
   }
 }
